+++ src/main/java/com/sparta/greeypeople/like/menu/MenuLikes.java
@@ -1,11 +1,10 @@
 package com.sparta.greeypeople.like.menu;
 
 import com.sparta.greeypeople.menu.entity.Menu;
-import com.sparta.greeypeople.timestamp.TimeStamp;
+import com.sparta.greeypeople.common.TimeStamp;
 
 import com.sparta.greeypeople.user.entity.User;
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
