+++ src/main/java/com/sparta/greeypeople/like/review/ReviewLikes.java
@@ -1,11 +1,10 @@
 package com.sparta.greeypeople.like.review;
 
 import com.sparta.greeypeople.review.entity.Review;
-import com.sparta.greeypeople.timestamp.TimeStamp;
+import com.sparta.greeypeople.common.TimeStamp;
 
 import com.sparta.greeypeople.user.entity.User;
 import jakarta.persistence.*;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
