+++ src/main/java/com/sparta/greeypeople/order/entity/Order.java
@@ -3,7 +3,7 @@ package com.sparta.greeypeople.order.entity;
 import com.sparta.greeypeople.menu.entity.Menu;
 import com.sparta.greeypeople.order.enumeration.Process;
 import com.sparta.greeypeople.store.entity.Store;
-import com.sparta.greeypeople.timestamp.TimeStamp;
+import com.sparta.greeypeople.common.TimeStamp;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
