+++ src/main/java/com/sparta/greeypeople/review/entity/Review.java
@@ -2,7 +2,7 @@ package com.sparta.greeypeople.review.entity;
 
 import com.sparta.greeypeople.review.dto.request.ReviewRequestDto;
 import com.sparta.greeypeople.store.entity.Store;
-import com.sparta.greeypeople.timestamp.TimeStamp;
+import com.sparta.greeypeople.common.TimeStamp;
 import com.sparta.greeypeople.user.entity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
